UniHub/requirements.txt
django
mysqlclient
mysql-connector-python
djangorestframework
djangorestframework-simplejwt
sendgrid
pyotp
Pillow
python-dotenv
fastapi
uvicorn
requests